backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-braces==1.12.0     # via django-model-controller
django-bulk-update==2.2.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.1  # via django-model-controller
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.0
django-filter==1.1.0
django-model-controller==0.1.0
django-reversion==2.0.13
django==2.0.1
djangorestframework-jwt==1.11.0
djangorestframework==3.7.7
gevent==1.2.2
greenlet==0.4.13          # via gevent
gunicorn==19.7.1
psycopg2==2.7.4
pyjwt==1.6.0              # via djangorestframework-jwt
pytz==2018.3              # via django
six==1.11.0               # via django-environ, django-extensions
sqlparse==0.2.4           # via django-debug-toolbar
typing==3.6.4             # via django-extensions
django-safedelete==0.4.5